Review common deploy logic
